### 2025-06-20 #8440 @editorialbot
[PRE REVIEW]: H-HIGNN Toolkit: A Software for Efficient and Scalable Simulation of Large-Scale Particulate Suspensions Using GNNs and H-Matrices
**Submitting author:** @zishengye (<a href="http://orcid.org/0000-0001-6675-9747">Zisheng Ye</a>)
**Repository:** https://github.com/Pan-Group-UW-Madison/hignn
**Branch with paper.md** (empty if default branch): joss-doc
**Version:** v1.0
