### 2025-07-22 #8440 @crvernon
:wave: @jbytecode are you able to take this one on as editor?
