### 2025-07-22 #8440 @jbytecode
@crvernon - sure!
