### 2025-07-28 #8440 @jbytecode
@Wenxiao-Pan - Of course I can invite them by using their e-mails but I can not force them to create a GitHub account.
