### 2025-08-02 #8440 @XiaozheHu
Dear Professor Mehmet Hakan Satman,

Thanks for the invitation and yes, I am happy to review the submission.

Best regards,
Xiaozhe
