### 2025-08-06 #8440 @jbytecode
👋👋👋 Dear @hsinyilin19 👋👋👋

**[[PRE REVIEW]: H-HIGNN Toolkit: A Software for Efficient and Scalable Simulation of Large-Scale Particulate Suspensions Using GNNs and H-Matrices](https://github.com/openjournals/joss-reviews/issues/8440)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/8440)).

Thank you in advance!
