### 2025-08-08 #8440 @hsinyilin19
Hi Professor Mehmet Hakan Satman,

Thank you for the kind invitation. Even though the project looks
interesting and powerful, it falls further from my area of expertise, so I
don’t feel qualified to provide a thorough review. I hope you secure
knowledgeable reviewers quickly and wish you all the best with the process.

Best regards,
Hsin-Yi Lin

On Wed, Aug 6, 2025 at 4:17 PM Mehmet Hakan Satman ***@***.***>
wrote:

> *jbytecode* left a comment (openjournals/joss-reviews#8440)
> <https://github.com/openjournals/joss-reviews/issues/8440#issuecomment-3161469681>
>
> 👋👋👋 Dear @hsinyilin19 <https://github.com/hsinyilin19> 👋👋👋
>
> Would you be willing to assist in reviewing this submission for *JOSS* (*Journal
> of Open Source Software*)?
>
> JOSS publishes articles about open source research software. The
> submission I'd like you to review is titled:
>
> *[PRE REVIEW]: H-HIGNN Toolkit: A Software for Efficient and Scalable
> Simulation of Large-Scale Particulate Suspensions Using GNNs and H-Matrices
> <https://github.com/openjournals/joss-reviews/issues/8440>*
>
> You can find more information at the top of this Github issue (LINK
> <https://github.com/openjournals/joss-reviews/issues/8440>).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is
> open, and author-reviewer-editor conversations are encouraged. If you have
> any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will
> start the review process in a separate thread. In that thread, there will
> be 23 check items for each single reviewer.
>
> Thank you in advance!
>
