### 2025-08-11 #8777 @editorialbot
[REVIEW]: H-HIGNN Toolkit: A Software for Efficient and Scalable Simulation of Large-Scale Particulate Suspensions Using GNNs and H-Matrices
**Submitting author:** @zishengye (<a href="http://orcid.org/0000-0001-6675-9747">Zisheng Ye</a>)
**Repository:** https://github.com/Pan-Group-UW-Madison/hignn
**Branch with paper.md** (empty if default branch): joss-doc
**Version:** v1.0
