### 2025-08-11 #8777 @jbytecode
Dear Reviewers @XiaozheHu and @rajeshrinet,

You can start by creating your task lists. Each list will contain several tasks.

As you complete each task, you can check off the corresponding checkbox. Since the review process for JOSS is interactive, you are encouraged to engage with the author, other reviewers, and the editor throughout. You can open issues and submit pull requests in the target repository. Please include the URL of this page in those interactions, so we can keep track of activities outside of the platform.

To generate your task list, simply type:

`@editorialbot generate my checklist`

Here is the review guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Thank you in advance.
