### 2025-08-19 #8777 @rajeshrinet
@jbytecode thanks for the reminder. I've started the review and am working my way through the checklist. I am adding some comments for the authors to consider. I'll continue with the rest of the review and will be in touch with further comments.

I've read the JOSS paper and found it to be a concise. This is an impressive and timely work. I could gather the scientific context and the practical application of H-HIGNN  by reading a paper on arXiv: https://arxiv.org/abs/2505.08174v2.

### Enhancing the 'State of the Field' Section
As per the review checklist, the "State of the field" section of the paper could be more complete. The software's capabilities, particularly concerning the computation of mobility matrices, could be put in better context by mentioning and briefly comparing them to other existing libraries.

I suggest that the author consider including a discussion of the following libraries, which are known for similar functionality :

* Hydrolib: https://doi.org/10.1016/0010-4655(95)00029-F

* libstokes: https://github.com/kichiki/libstokes

* Stokesdt: https://github.com/xing-liu/stokesdt

* PyStokes: https://github.com/rajeshrinet/pystokes

As a full disclosure, I am involved in the development of PyStokes.

### Community guidelines

#### Contributing Guidelines
A CONTRIBUTING.md file is missing. Such a file is vital component for an open-source project. This document provides clear instructions for how others can contribute, from reporting bugs to submitting code.

#### Code of Conduct:

Needs to be added, which is essential for fostering a positive and inclusive community.

## Example usage

The documentation of the code can be improved for new users and potential contributors. For example, currently users are not guided to any example files in the repository through the documentation or the README.

### Jupyter notebook examples
The authors can consider this suggestion as optional. It is helpful to have some example files in the format of jupyter notebooks (see https://jupyter.org/).

In particular, one can explain a bit of maths, run the code, and compare results to existing literature - all in the same window! Adding such notebooks will increase the reach of the software and make it more user-friendly.
