### 2025-09-20 #8777 @jbytecode
@oak97, @jonaspleyer - We already assigned a reviewer here but the other reviewer has a COI. I need one more suitable reviewer for this submission. Would you like to assist in reviewing this submission for JOSS? Thank you in advance.
