### 2025-09-20 #8777 @jonaspleyer
Hi @jbytecode thanks for reaching out. I would be happy to take on the review. I will probably be able to get started within the next week.
