### 2025-09-20 #8777 @jbytecode
@marcobonici - please start your review by typing `@editorialbot generate my checklist`

Thank you in advance
