### 2025-09-25 Pan-Group-UW-Madison/hignn#10 @jonaspleyer
[JOSS-REVIEW] Problems with preparing the docker environment
When running the instructions on the `README.md` file to prepare the docker environment, I end up with this error:

```
$ docker build --rm -f Dockerfile.hignn.Ampreme -t hignn .

DEPRECATED: The legacy builder is deprecated and will be removed in a future release.
            Install the buildx component to build images with BuildKit:
            https://docs.docker.com/go/buildx/

Sending build context to Docker daemon  33.79kB
Step 1/56 : FROM nvidia/cuda:11.8.0-cudnn8-devel-ubuntu20.04
unauthorized: authentication required
```

| | |
| --- | --- |
| Docker | 28.4.0, build d8eb465f86 |
| cuda | 12.9.1-2 |
| nvidia-container-toolkit | 1.17.8-1 |

I am not sure but do I need to be logged in to docker hub?

I get a similar error message when trying any of the other commands to set up the docker environment.

```
$ docker pull panlabuwmadison/hignn:latest
Error response from daemon: unauthorized: authentication required
```

https://github.com/openjournals/joss-reviews/issues/8777
