### 2025-09-25 Pan-Group-UW-Madison/hignn#12 @jonaspleyer
[JOSS-REVIEW] (Automated) Tests
I did find the tests for the repository in the `test` folder. It seems that they are not run automatically but need to be applied manually. Consider using a github actions file under `.github/workflows/test.yml` that runs the tests for every change within the repository.
See: https://docs.github.com/en/actions

https://github.com/openjournals/joss-reviews/issues/8777
