### 2025-10-02 Pan-Group-UW-Madison/hignn#13 @jonaspleyer
[JOSS-REVIEW] Missing Community Guidelines
The submission guidelines for JOSS state that community guidelines and means of contributing should be present within the repository. This seems to be missing so far.

https://github.com/openjournals/joss-reviews/issues/8777
