### 2025-10-02 Pan-Group-UW-Madison/hignn#14 @jonaspleyer
[JOSS-REVIEW] Missing Documentation
The `README.md` file contains some useful information to get started with the project. However a dedicated documentation seems to be missing.

https://github.com/openjournals/joss-reviews/issues/8777
