### 2025-11-01 #8777 @jbytecode
@jonaspleyer - Thank you for your detailed work and explanation.

The JOSS article clearly appears related to both the pre-print and the article published in the Journal of Computational Physics. I don't think it's right for us to restrict ourselves regarding citations. I suggest that the authors also cite the article in the physics journal and highlight any differences if they exist. If they are not related, it should be carefully explained why they bear the same name. In this situation, citing seems more appropriate than not citing.
