### 2025-11-01 #8777 @rajeshrinet
@jonaspleyer @jbytecode and @Wenxiao-Pan I have been closely following the very interesting discussion.

I would like to add the I have indicated some related references in my comments above: https://github.com/openjournals/joss-reviews/issues/8777#issuecomment-3199529760

I have indicated four software which exists. These softwares are particle-based and include hydrodynamic interactions between particles. They have been used to publish papers in referred journals. Some of these also present prescriptions to add activity and noise. Thus, these softwares are quite relevant to the current study.

Apart from that I have suggested changes to documentation and community guidelines, which have also been indicated by @jonaspleyer.
