### 2025-11-15 #8777 @Wenxiao-Pan
@jbytecode I will address the revisions on the paper related to citations and discussions on related software. How would you like us to submit the revised paper?
