### 2025-11-15 Pan-Group-UW-Madison/hignn#10 @jonaspleyer
The docker container runs fine for now. However I get compilation errors:

```
/local/src/FarDot.cpp: In lambda function:
/local/src/FarDot.cpp:1104:44: error: 'min' was not declared in this scope
 1104 |                 workingNodeIteration(i) >= min(rowSize, colSize)) {
      |                                            ^~~
/local/src/FarDot.cpp:1104:44: note: suggested alternatives:
In file included from /usr/include/c++/9/algorithm:62,
                 from /local/include/HignnModel.hpp:4,
                 from /local/src/FarDot.cpp:1:
/usr/include/c++/9/bits/stl_algo.h:3456:5: note:   'std::min'
 3456 |     min(initializer_list<_Tp> __l, _Compare __comp)
      |     ^~~
In file included from /opt/pytorch/torch/include/ATen/Functions.h:956,
                 from /opt/pytorch/torch/include/ATen/ATen.h:13,
                 from /opt/pytorch/torch/include/torch/csrc/api/include/torch/types.h:3,
                 from /opt/pytorch/torch/include/torch/script.h:3,
                 from /local/include/HignnModel.hpp:18,
                 from /local/src/FarDot.cpp:1:
/opt/pytorch/torch/include/ATen/ops/min.h:77:19: note:   'at::min'
   77 | inline at::Tensor min(const at::Tensor & self, const at::Tensor & other) {
      |                   ^~~
In file included from /opt/kokkos/include/View/Kokkos_ViewLegacy.hpp:44,
                 from /opt/kokkos/include/Kokkos_View.hpp:29,
                 from /opt/kokkos/include/Kokkos_Parallel.hpp:31,
                 from /opt/kokkos/include/OpenMP/Kokkos_OpenMP.hpp:32,
                 from /opt/kokkos/include/decl/Kokkos_Declare_OPENMP.hpp:21,
                 from /opt/kokkos/include/KokkosCore_Config_DeclareBackend.hpp:22,
                 from /opt/kokkos/include/Kokkos_Core.hpp:45,
                 from /local/include/Typedef.hpp:12,
                 from /local/include/HignnModel.hpp:28,
                 from /local/src/FarDot.cpp:1:
/opt/kokkos/include/Kokkos_MinMax.hpp:89:36: note:   'Kokkos::min'
   89 | KOKKOS_INLINE_FUNCTION constexpr T min(std::initializer_list<T> ilist,
      |                                    ^~~
In file included from /opt/pytorch/torch/include/torch/csrc/jit/ir/ir.h:18,
                 from /opt/pytorch/torch/include/torch/csrc/jit/api/function_impl.h:4,
                 from /opt/pytorch/torch/include/torch/csrc/jit/api/method.h:7,
                 from /opt/pytorch/torch/include/torch/csrc/jit/api/object.h:5,
                 from /opt/pytorch/torch/include/torch/csrc/jit/api/module.h:4,
                 from /opt/pytorch/torch/include/torch/csrc/jit/serialization/import.h:5,
                 from /opt/pytorch/torch/include/torch/script.h:9,
                 from /local/include/HignnModel.hpp:18,
                 from /local/src/FarDot.cpp:1:
/opt/pytorch/torch/include/ATen/core/interned_strings.h:352:1: note:   'c10::prim::min'
  352 | FORALL_NS_SYMBOLS(DEFINE_SYMBOL)
      | ^~~~~~~~~~~~~~~~~
/opt/pytorch/torch/include/ATen/core/interned_strings.h:352:1: note:   'c10::aten::min'
  352 | FORALL_NS_SYMBOLS(DEFINE_SYMBOL)
      | ^~~~~~~~~~~~~~~~~
/opt/pytorch/torch/include/ATen/core/interned_strings.h:352:1: note:   'c10::attr::min'
  352 | FORALL_NS_SYMBOLS(DEFINE_SYMBOL)
      | ^~~~~~~~~~~~~~~~~
In file included from /opt/pytorch/torch/include/ATen/MethodOperators.h:268,
                 from /opt/pytorch/torch/include/ATen/core/TensorBody.h:41,
                 from /opt/pytorch/torch/include/ATen/core/ivalue.h:4,
                 from /opt/pytorch/torch/include/ATen/core/List_inl.h:4,
                 from /opt/pytorch/torch/include/ATen/core/List.h:488,
                 from /opt/pytorch/torch/include/ATen/core/IListRef_inl.h:3,
                 from /opt/pytorch/torch/include/ATen/core/IListRef.h:631,
                 from /opt/pytorch/torch/include/ATen/DeviceGuard.h:3,
                 from /opt/pytorch/torch/include/ATen/ATen.h:9,
                 from /opt/pytorch/torch/include/torch/csrc/api/include/torch/types.h:3,
                 from /opt/pytorch/torch/include/torch/script.h:3,
                 from /local/include/HignnModel.hpp:18,
                 from /local/src/FarDot.cpp:1:
/opt/pytorch/torch/include/ATen/ops/min_ops.h:61:18: note:   'at::_ops::min'
   61 | struct TORCH_API min {
      |                  ^~~
In file included from /opt/pytorch/torch/include/ATen/NativeFunctions.h:910,
                 from /opt/pytorch/torch/include/ATen/TensorIndexing.h:13,
                 from /opt/pytorch/torch/include/ATen/ATen.h:18,
                 from /opt/pytorch/torch/include/torch/csrc/api/include/torch/types.h:3,
                 from /opt/pytorch/torch/include/torch/script.h:3,
                 from /local/include/HignnModel.hpp:18,
                 from /local/src/FarDot.cpp:1:
/opt/pytorch/torch/include/ATen/ops/min_native.h:29:22: note:   'at::native::min'
   29 | TORCH_API at::Tensor min(const at::Tensor & self, const at::Tensor & other);
      |                      ^~~
In file included from /local/extern/eigen/Eigen/Core:191,
                 from /local/include/HignnModel.hpp:22,
                 from /local/src/FarDot.cpp:1:
/local/extern/eigen/Eigen/src/Core/arch/Default/Half.h:736:44: note:   'Eigen::half_impl::min'
  736 | EIGEN_STRONG_INLINE EIGEN_DEVICE_FUNC half(min)(const half& a, const half& b) {
      |                                            ^~~
In file included from /local/extern/eigen/Eigen/Core:192,
                 from /local/include/HignnModel.hpp:22,
                 from /local/src/FarDot.cpp:1:
/local/extern/eigen/Eigen/src/Core/arch/Default/BFloat16.h:651:48: note:   'Eigen::bfloat16_impl::min'
  651 | EIGEN_STRONG_INLINE EIGEN_DEVICE_FUNC bfloat16(min)(const bfloat16& a, const bfloat16& b) {
      |                                                ^~~
make[2]: *** [CMakeFiles/hignn.dir/build.make:135: CMakeFiles/hignn.dir/src/FarDot.cpp.o] Error 1
make[2]: *** Waiting for unfinished jobs....
make[1]: *** [CMakeFiles/Makefile2:109: CMakeFiles/hignn.dir/all] Error 2
make: *** [Makefile:91: all] Error 2
cp: cannot stat '*.so': No such file or directory
cp: cannot stat '*.so': No such file or directory
```

It seems that modifying line `1104` from `min` to `std::min` fixes the error.
