+++ paper/paper.md
@@ -1,0 +1,73 @@
+---
+title: '$\mathcal{H}$-HIGNN Toolkit: A Software for Efficient and Scalable Simulation of Large-Scale Particulate Suspensions Using GNNs and $\mathcal{H}$-Matrices'
+
+tags: 
+  - Python/C++
+  - Complex Fluids
+  - Fluid Dynamics
+  - Graph Neural Networks 
+  - Hierarchical Matrices
+
+authors:
+  - name: Zisheng Ye
+    orcid: 0000-0001-6675-9747
+    affiliation: 1
+
+  - name: Zhan Ma
+    affiliation: 1
+
+  - name: Ebrahim Safdarian
+    affiliation: 1
+
+  - name: Shirindokht Yazdani
+    affiliation: 1
+
+  - name: Wenxiao Pan
+    orcid: 0000-0002-2252-7888
+    corresponding: true
+    affiliation: 1
+
+affiliations:
+  - name: Department of Mechanical Engineering, University of Wisconsin-Madison, Madison, WI, USA 
+    index: 1
+
+bibliography: paper.bib
+
+---
+
+# Summary
+
+Particulate suspensions—systems of particles dispersed in viscous fluids—play a critical role in various scientific and engineering applications [@ParticulateReview_Maxey2017,@shelley2016dynamics]. This software implements $\mathcal{H}$-HIGNN, a framework designed for efficient and scalable simulation of large-scale particulate suspensions (Figure 1). It extends the Hydrodynamic Interaction Graph Neural Network (HIGNN) approach [@ma2022fast, @ma2024shape], which utilizes GNNs to model the mobility tensor that dictates particle dynamics under hydrodynamic interactions (HIs) and external forces. HIGNN effectively captures both short- and long-range HIs and their many-body effects and enables substantial computational acceleration by harvesting the power of machine learning. By incorporating hierarchical matrix ($\mathcal{H}$-matrix) techniques, $\mathcal{H}$-HIGNN further improves computational efficiency, achieving quasi-linear prediction cost with respect to the number of particles. Its GPU-optimized implementation delivers near-theoretical quasi-linear wall-time scaling and near-ideal strong scalability for parallel efficiency. The methodology, validation, and efficiency demonstrations of $\mathcal{H}$-HIGNN are detailed in [@ma2025].
+
+![ (Left) Scalability test of wall time with respect to $N$ (number of particles) and $N_\text{GPU}$ (number of GPUs). (Right) Example large-scale particulate suspensions simulated by $\mathcal{H}$-HIGNN Toolkit.\label{fig:illustration}](assets/llustration.png)
+
+# Statement of need
+
+Simulating particulate suspensions in 3D poses substantial computational challenges, limiting prior work to small numbers of particles or requiring extensive computational resources. This emphasizes the need for an efficient, scalable, and flexible toolkit that enables researchers to investigate practically relevant, large-scale suspensions, pushing the boundaries beyond previously accessible scales while minimizing computational resource demands. The present software addresses this gap by introducing the first linearly scalable toolkit capable of simulating suspensions with millions of particles or more using only modest resources, such as a few mid-range GPUs [@ma2025]. Beyond rigid passive particles, the $\mathcal{H}$-HIGNN toolkit is flexible to be extended to simulate suspensions of soft matter systems, such as flexible filaments or membranes, through the inclusion of additional interparticle interaction forces, and to support the simulation of active matter, such as microswimmers, by incorporating active forces or actuation fields. Therefore, this software offers a powerful platform for exploring hydrodynamic effects across a broad range of systems in soft and active matter.
+
+# Description of the software
+
[…]
