+++ paper/paper.bib
@@ -1,0 +1,105 @@
+@article{townsend2024stokesian,
+    title = {Stokesian dynamics in Python},
+    author={Townsend, Adam K},
+    journal={Journal of Open Source Software},
+    volume={9},
+    number={94},
+    pages={6011},
+    year={2024}
+}
+
+@article{torre2025python,
+    title={Python-JAX-based fast Stokesian dynamics},
+    author={Torre, Kim William and Schram, Raoul D and de Graaf, Joost},
+    journal={arXiv preprint arXiv:2503.07847},
+    year={2025}
+}
+
+@article{ma2025,
+    title={$\mathcal{H}$-HIGNN: A Scalable Graph Neural Network Framework with Hierarchical Matrix Acceleration for Simulation of Large-Scale Particulate Suspensions},
+    author={Zhan Ma and Zisheng Ye and Ebrahim Safdarian and Wenxiao Pan},
+    journal={Journal of Computational Physics},
+    volume={544},
+    pages={114429},
+    year={2026}
+}
+
+@article{ma2022fast,
+    title={Fast simulation of particulate suspensions enabled by graph neural network},
+    author={Ma, Zhan and Ye, Zisheng and Pan, Wenxiao},
+    journal={Computer Methods in Applied Mechanics and Engineering},
+    volume={400},
+    pages={115496},
+    year={2022}
+}
+
+@article{ma2024shape,
+    title={Shape deformation, disintegration, and coalescence of suspension drops: {E}fficient simulation enabled by graph neural networks},
+    author={Ma, Zhan and Pan, Wenxiao},
+    journal={International Journal of Multiphase Flow},
+    volume={176},
+    pages={104845},
+    year={2024}
+}
+
+@article{shelley2016dynamics,
+    title={The dynamics of microtubule/motor-protein assemblies in biology and physics},
+    author={Shelley, Michael J},
+    journal={Annual review of fluid mechanics},
+    volume={48},
+    number={1},
+    pages={487--506},
+    year={2016}
+}
+
+@article{ParticulateReview_Maxey2017,
+    author = {Maxey, Martin},
+    title = {Simulation Methods for Particulate Flows and Concentrated Suspensions},
+    journal = {Annual Review of Fluid Mechanics},
+    volume = {49},
+    number = {1},
+    pages = {171-193},
+    year = {2017}
+}
+
+@article{singh2019pystokes,
+    title={Pystokes: Phoresis and Stokesian hydrodynamics in python},
+    author={Singh, Rajesh and Adhikari, Ronojoy},
+    journal={arXiv preprint arXiv:1910.00909},
+    year={2019}
+}
+
+@misc{Stokesdt,
+    title={StokesDT},
+    url={https://github.com/xing-liu/stokesdt/wiki},
+    year={2015}
+}
+
+@inproceedings{liu2014large,
+    title={Large-scale hydrodynamic Brownian simulations on multicore and manycore architectures},
+    author={Liu, Xing and Chow, Edmond},
+    booktitle={2014 IEEE 28th International Parallel and Distributed Processing Symposium},
+    pages={563--572},
+    year={2014},
+    organization={IEEE}
+}
+
+@article{hinsen1995hydrolib,
+    title={HYDROLIB: a library for the evaluation of hydrodynamic interactions in colloidal suspensions},
+    author={Hinsen, K},
+    journal={Computer physics communications},
+    volume={88},
+    number={2-3},
+    pages={327--340},
+    year={1995}
+}
+
+@article{olsen2023openfoam,
+    title={An OpenFOAM solver for computing suspended particles in water currents},
[…]
