+++ paper/paper.bib
@@ -5,14 +5,16 @@
     volume={9},
     number={94},
     pages={6011},
-    year={2024}
+    year={2024},
+    doi={10.21105/joss.06011}
 }
 
 @article{torre2025python,
     title={Python-JAX-based fast Stokesian dynamics},
     author={Torre, Kim William and Schram, Raoul D and de Graaf, Joost},
     journal={arXiv preprint arXiv:2503.07847},
-    year={2025}
+    year={2025},
+    doi={10.21468/scipostphyscodeb.56}
 }
 
 @article{ma2025,
@@ -21,7 +23,8 @@
     journal={Journal of Computational Physics},
     volume={544},
     pages={114429},
-    year={2026}
+    year={2026},
+    doi={10.1016/j.jcp.2025.114429}
 }
 
 @article{ma2022fast,
@@ -30,7 +33,8 @@
     journal={Computer Methods in Applied Mechanics and Engineering},
     volume={400},
     pages={115496},
-    year={2022}
+    year={2022},
+    doi={10.1016/j.cma.2022.115496}
 }
 
 @article{ma2024shape,
@@ -39,7 +43,8 @@
     journal={International Journal of Multiphase Flow},
     volume={176},
     pages={104845},
-    year={2024}
+    year={2024},
+    doi={10.1016/j.ijmultiphaseflow.2024.104845}
 }
 
 @article{shelley2016dynamics,
@@ -49,7 +54,8 @@
     volume={48},
     number={1},
     pages={487--506},
-    year={2016}
+    year={2016},
+    doi={10.1146/annurev-fluid-010814-013639}
 }
 
 @article{ParticulateReview_Maxey2017,
@@ -59,14 +65,21 @@
     volume = {49},
     number = {1},
     pages = {171-193},
-    year = {2017}
+    year = {2017},
+    doi = {10.1146/annurev-fluid-122414-034408}
 }
 
 @article{singh2019pystokes,
-    title={Pystokes: Phoresis and Stokesian hydrodynamics in python},
-    author={Singh, Rajesh and Adhikari, Ronojoy},
-    journal={arXiv preprint arXiv:1910.00909},
-    year={2019}
+    doi = {10.21105/joss.02318},
+    url = {https://doi.org/10.21105/joss.02318},
+    year = {2020},
+    publisher = {The Open Journal},
+    volume = {5},
+    number = {50},
+    pages = {2318},
+    author = {Singh, Rajesh and Adhikari, Ronojoy},
+    title = {PyStokes: phoresis and Stokesian hydrodynamics in Python},
+    journal = {Journal of Open Source Software}
 }
 
 @misc{Stokesdt,
@@ -81,7 +94,8 @@
     booktitle={2014 IEEE 28th International Parallel and Distributed Processing Symposium},
     pages={563--572},
     year={2014},
-    organization={IEEE}
+    organization={IEEE},
+    doi={10.1109/ipdps.2014.65}
 }
 
 @article{hinsen1995hydrolib,
@@ -91,7 +105,8 @@
     volume={88},
     number={2-3},
     pages={327--340},
-    year={1995}
+    year={1995},
+    doi={10.1016/0010-4655(95)00029-f}
 }
 
 @article{olsen2023openfoam,
@@ -101,5 +116,6 @@
     volume={25},
     number={5},
     pages={1949--1959},
-    year={2023}
+    year={2023},
+    doi={10.2166/hydro.2023.309}
 }
