### 2025-11-26 #8777 @zishengye
@jbytecode I've added all suggested DOIs.
