### 2025-11-30 Pan-Group-UW-Madison/hignn#12 @zishengye
@jonaspleyer I've added automatic test in the main branch.
