### 2025-12-18 Pan-Group-UW-Madison/hignn#14 @jonaspleyer
The generated documentation is very nice. I have successfully executed the `doxygen` command as described. However I think that it would still be worthwhile to host this documentation i.e. on Github pages via a simple actions script. I find the actual documentation of the API methods satisfactory to my liking.
