### 2025-12-18 37830e9
Add GitHub Actions workflow for Doxygen docs deployment
