+++ .github/workflows/deploy-docs.yml
@@ -1,0 +1,35 @@
+name: Deploy Doxygen Docs to Pages
+
+on:
+  push:
+    branches: [ "main" ] # 在推送到 main 分支时触发
+  workflow_dispatch: # 允许手动触发
+
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
+jobs:
+  build-and-deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
+
+      - name: Install Doxygen
+        run: sudo apt-get update && sudo apt-get install -y doxygen graphviz
+
+      - name: Generate Doxygen HTML
+        run: doxygen Doxyfile # 使用你已有的配置文件
+
+      - name: Setup Pages
+        uses: actions/configure-pages@v4
+
+      - name: Upload artifact
+        uses: actions/upload-pages-artifact@v3
+        with:
+          path: './html' # 必须与 HTML_OUTPUT 设置的路径一致
+
+      - name: Deploy to GitHub Pages
+        uses: actions/deploy-pages@v4
