### 2025-12-18 Pan-Group-UW-Madison/hignn#14 @zishengye
Thanks for your suggestion. I have set the GitHub Pages and you can view it [here](https://pan-group-uw-madison.github.io/hignn/).

@Wenxiao-Pan You can check the documents online now.
