### 2025-12-18 #8777 @marcobonici
The paper is nice, a few remarks:
1. There is a typo, it is written **visuliaze.py** rather than **visualize.py** (line 60).
2. I see the usage of "—", which often indicates the usage of LLM systems for writing. There is nothing wrong with it, but if that is the case, a short statement about it would be ideal.
3. Figure 1 seems to be a png, rather than a vectorized image. Ideally, this should be an embedded pdf, so that the quality of the image is not degraded upon zoom.
4. Why is there a reference to a paper (Ma et al. 2026) from the future?
5. Also, as @jonaspleyer noticed, some additional citations could be added to the "Statement of need" section.
