### 2025-12-18 #8777 @zishengye
@marcobonici
Thanks for your suggestions. We will follow your remarks for corrections and revise our paper.

As for
> 4. Why is there a reference to a paper (Ma et al. 2026) from the future?

[This article](https://doi.org/10.1016/j.jcp.2025.114429) has already been formally accepted for publication in Journal of Computational Physics. The listed publication year of 2026 refers to the assigned volume and issue for the final, typeset version in the journal's scheduled online issue.
