### 2025-12-27 #8777 @zishengye
@jbytecode I am close to the last step but I am curious how can I modify dpi of the figures included in the manuscript? The first figure in the manuscript was merged from four different figures. One of them is originally png. So, I can't make them all into pdf. But the original figure quality in the asset folder is very high but the generation process has greatly compressed the figure. Do you any proper way to preserve the original png quality to make the figure clear enough.
