### 2026-01-04 ca7a5f1
minor fixes on manuscript and the bibliography
