+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{townsend2024stokesian,
-    title = {Stokesian dynamics in Python},
+    title = {Stokesian dynamics in {P}ython},
     author={Townsend, Adam K},
     journal={Journal of Open Source Software},
     volume={9},
@@ -10,7 +10,7 @@
 }
 
 @article{torre2025python,
-    title={Python-JAX-based fast Stokesian dynamics},
+    title={Python-{JAX}-based fast {S}tokesian dynamics},
     author={Torre, Kim William and Schram, Raoul D and de Graaf, Joost},
     journal={arXiv preprint arXiv:2503.07847},
     year={2025},
@@ -78,7 +78,7 @@
     number = {50},
     pages = {2318},
     author = {Singh, Rajesh and Adhikari, Ronojoy},
-    title = {PyStokes: phoresis and Stokesian hydrodynamics in Python},
+    title = {PyStokes: phoresis and {S}tokesian hydrodynamics in {P}ython},
     journal = {Journal of Open Source Software}
 }
 
@@ -89,7 +89,7 @@
 }
 
 @inproceedings{liu2014large,
-    title={Large-scale hydrodynamic Brownian simulations on multicore and manycore architectures},
+    title={Large-scale hydrodynamic {B}rownian simulations on multicore and manycore architectures},
     author={Liu, Xing and Chow, Edmond},
     booktitle={2014 IEEE 28th International Parallel and Distributed Processing Symposium},
     pages={563--572},
