### 2026-01-18 #8777 @jbytecode
@zishengye, @Wenxiao-Pan - Please create a tagged release in the repo, use this tagged release to create a Zenodo archive. You can use the zipped (or tar.gz'ed) mirror by uploading it in the Zenodo archive.
