### 2026-01-19 #8777 @Wenxiao-Pan
@jbytecode I manually upload the zipped file to Zenodo and obtain this DOI badge: https://doi.org/10.5281/zenodo.18302253
