### 2026-01-19 #8777 @jbytecode
@zishengye, @Wenxiao-Pan - Thank you for creating the archive. We have some issues:

- The author list and their orcids should match those in the archive.
- The license of software should match that in the archive.

Please fix and then ping me. Thank you in advance.
