### 2026-01-20 #8777 @Wenxiao-Pan
> [@zishengye](https://github.com/zishengye), [@Wenxiao-Pan](https://github.com/Wenxiao-Pan) - Thank you for creating the archive. We have some issues:
>
> * The author list and their orcids should match those in the archive.
> * The license of software should match that in the archive.
>
> Please fix and then ping me. Thank you in advance.

@jbytecode I think I have fixed these issues. Please check. Thanks!
