### 2026-01-20 #8777 @jbytecode
@Wenxiao-Pan - Sorry, we don't need to specify who are the supervisors and who are the project members.

In the archive, the author is Pan Group@UW-Madison, but the paper has multiple authors. Please list the authors and their metadata here. The author list and their metadata in the manuscript should match those in the archive.

Thank you in advance.
