### 2026-01-21 #8777 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
