### 2026-01-23 #8777 @crvernon
@zishengye - I just need you to address the following before I accept:

In your paper:

- [x] LINE 21:  "...detailed in (Ma et al., 2026)" should be cited in-text as "...detailed in Ma et al. (2026)" this should be just a minor bracket alteration in your paper file.

That's it.  Everything else looks great.  No new release is required after this change has been made.  Let me know when it is done.  Thank you.
