### 2026-01-25 1c823db
fix citation format
