### 2026-01-26 #8777 @crvernon
@zishengye I just noticed there is special formatting applied to your title in your paper.  Please remove any special formatting as it messes up the way the paper title shows when it is published in JOSS.

Let me know when this has been done and I'll reaccept this paper to ensure that it appears correctly.
