### 2026-01-26 deab6fd
remove special chars in title
