### 2026-01-26 #8777 @crvernon
🥳 Congratulations on your new publication @zishengye ! Many thanks to @jbytecode for editing and @rajeshrinet, @jonaspleyer, and @marcobonici for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
